requirements.txt
pandas
psycopg2-binary
duckdb
litellm
python-dotenv
requests